Changed the float to decimal, more precise. simplified price decision for bot company, increased capital for company and corrected buying bug

+++ Simul/Models/Calculator.cs
@@ -4,7 +4,7 @@ namespace Simul.Models
 {
     public static class Calculator
     {
-        public static float CalculateSkillIncrement(float skill)
+        public static decimal CalculateSkillIncrement(decimal skill)
         {
             if (skill <= 1)
             {
@@ -12,25 +12,25 @@ namespace Simul.Models
             }
             else if (skill <= 5)
             {
-                skill += 0.5f;
+                skill += 0.5m;
             }
             else if (skill <= 10)
             {
-                skill += 0.25f;
+                skill += 0.25m;
             }
             else if (skill <= 20)
             {
-                skill += 0.10f;
+                skill += 0.10m;
             }
             else
             {
-                skill += 0.05f;
+                skill += 0.05m;
             }
 
             return skill;
         }
 
-        public static float CalculateStrengthIncrement(float strength)
+        public static decimal CalculateStrengthIncrement(decimal strength)
         {
             if (strength <= 10)
             {
@@ -38,25 +38,25 @@ namespace Simul.Models
             }
             else if (strength <= 50)
             {
-                strength += 0.5f;
+                strength += 0.5m;
             }
             else if (strength <= 100)
             {
-                strength += 0.25f;
+                strength += 0.25m;
             }
             else if (strength <= 200)
             {
-                strength += 0.10f;
+                strength += 0.10m;
             }
             else
             {
-                strength += 0.05f;
+                strength += 0.05m;
             }
 
             return strength;
         }
 
-        public static float CalculateProductionProgress(ICompany company, IPerson employee)
+        public static decimal CalculateProductionProgress(ICompany company, IPerson employee)
         {
             var skill = employee.Skillset.Skills.First(x => x.Key == company.ProducedResource.ImprovedSkill);
             return skill.Value;
